using fixed sized platform free int type

+++ v_filesystem/beanfs/beanfs.h
@@ -11,6 +11,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 #define BLOCK_SIZE 512
 // first inode is for root
@@ -22,48 +23,48 @@
 // using logical block address from 0 to user defined number
 struct beanfs_super_block {
     
-    unsigned long s_blocks_count;
-    unsigned long s_inodes_count;
-    unsigned long s_data_blocks_count;
-    unsigned long s_free_blocks_count;
-    unsigned long s_free_inodes_count;
+    uint32_t s_blocks_count;
+    uint32_t s_inodes_count;
+    uint32_t s_data_blocks_count;
+    uint32_t s_free_blocks_count;
+    uint32_t s_free_inodes_count;
     
-    unsigned long s_free_blocks_list_group[FREE_BLOCKS_LIST_SIZE];
-    unsigned long s_free_inodes_list_group[FREE_INODES_LIST_SIZE];
-    unsigned short blocks_list_top;              // number of blocks in free block list
-    unsigned short inodes_list_top;              // number of inodes in free inode list
+    uint32_t s_free_blocks_list_group[FREE_BLOCKS_LIST_SIZE];
+    uint32_t s_free_inodes_list_group[FREE_INODES_LIST_SIZE];
+    uint16_t blocks_list_top;              // number of blocks in free block list
+    uint16_t inodes_list_top;              // number of inodes in free inode list
     
-    unsigned long first_data_block;
-    unsigned long first_inode;
+    uint32_t first_data_block;
+    uint32_t first_inode;
     
-    //unsigned short block_list_lock;             //0 stand for unlock and 1 stand for unlock
-    //unsigned short inode_list_lock;
+    //uint16_t block_list_lock;             //0 stand for unlock and 1 stand for unlock
+    //uint16_t inode_list_lock;
     
-    unsigned long modify_time;                  // modified time stamp
+    uint32_t modify_time;                  // modified time stamp
 };
 
 // structure of superblock in memomry
 struct beanfs_sb_info {
     
     char i_ftype;
-    unsigned long s_blocks_count;
-    unsigned long s_inodes_count;
-    unsigned long s_free_blocks_count;
-    unsigned long s_free_inodes_count;
+    uint32_t s_blocks_count;
+    uint32_t s_inodes_count;
+    uint32_t s_free_blocks_count;
+    uint32_t s_free_inodes_count;
     
-    unsigned long s_free_blocks_list_group[60]; // array of blocks offset
-    unsigned long s_free_inodes_list_group[60]; // array of inodes offset
-    unsigned short blocks_list_top;              // number of blocks in free block list
-    unsigned short inodes_list_top;              // number of inodes in free inode list
+    uint32_t s_free_blocks_list_group[60]; // array of blocks offset
+    uint32_t s_free_inodes_list_group[60]; // array of inodes offset
+    uint16_t blocks_list_top;              // number of blocks in free block list
+    uint16_t inodes_list_top;              // number of inodes in free inode list
     
-    unsigned long first_data_block;             // first data block
-    unsigned long first_inode;                  // first inode
+    uint32_t first_data_block;             // first data block
+    uint32_t first_inode;                  // first inode
     
-    unsigned short blocks_list_lock;             //0 stand for unlock and 1 stand for unlock
-    unsigned short inodes_list_lock;
+    uint16_t blocks_list_lock;             //0 stand for unlock and 1 stand for unlock
+    uint16_t inodes_list_lock;
     
-    unsigned short is_modified;                 // flag for wheather modified
-    unsigned long modify_time;                  // modified time stamp
+    uint16_t is_modified;                 // flag for wheather modified
+    uint32_t modify_time;                  // modified time stamp
 };
 
 
@@ -74,22 +75,22 @@ struct beanfs_sb_info {
 struct beanfs_inode {
     
     char i_file_type;
-    unsigned short i_mode;                     // permission
-    unsigned short i_links;                    // file link count
-    unsigned long i_uid;                       // file onwer
-    unsigned long i_gid;                       // file owner group
+    uint16_t i_mode;                     // permission
+    uint16_t i_links;                    // file link count
+    uint32_t i_uid;                       // file onwer
+    uint32_t i_gid;                       // file owner group
     
-    unsigned long i_atime;                     // last access time
-    unsigned long i_mtime;                     // last modify time
-    unsigned long i_ctime;                     // last inode changetime
-    unsigned long i_birthtime;                 // creation time
+    uint32_t i_atime;                     // last access time
+    uint32_t i_mtime;                     // last modify time
+    uint32_t i_ctime;                     // last inode changetime
+    uint32_t i_birthtime;                 // creation time
     
-    unsigned long i_size;                      // file bytes count
-    unsigned long i_blocks;                    // actual blocks held
+    uint32_t i_size;                      // file bytes count
+    uint32_t i_blocks;                    // actual blocks held
     
     struct {
-        unsigned long i_db_offset[BEANFS_NDADDR];     // direct addr; number in data_blocks
-        unsigned long i_id_offset[BEANFS_NINDIR];     // first is 1st direct and second is 2th indirect
+        uint32_t i_db_offset[BEANFS_NDADDR];     // direct addr; number in data_blocks
+        uint32_t i_id_offset[BEANFS_NINDIR];     // first is 1st direct and second is 2th indirect
     } i_addr;
     
 };
@@ -97,40 +98,36 @@ struct beanfs_inode {
 // structure of inode in memory
 struct beanfs_inode_info {
     
-    unsigned short i_access_count;             // access count
+    uint16_t i_access_count;             // access count
     
-    unsigned short i_mode;                     // permission
-    unsigned short i_links;                    // file link count
-    unsigned long i_uid;                       // file onwer
-    unsigned long i_gid;                       // file owner group
+    uint16_t i_mode;                     // permission
+    uint16_t i_links;                    // file link count
+    uint32_t i_uid;                       // file onwer
+    uint32_t i_gid;                       // file owner group
     
-    unsigned long i_atime;                     // last access time
-    unsigned long i_mtime;                     // last modify time
-    unsigned long i_ctime;                     // last inode changetime
-    unsigned long i_birthtime;                 // creation time
+    uint32_t i_atime;                     // last access time
+    uint32_t i_mtime;                     // last modify time
+    uint32_t i_ctime;                     // last inode changetime
+    uint32_t i_birthtime;                 // creation time
     
-    unsigned long i_size;                      // file bytes count
-    unsigned long i_blocks;                    // actual blocks held
+    uint32_t i_size;                      // file bytes count
+    uint32_t i_blocks;                    // actual blocks held
     
     struct {
-        unsigned long i_db_offset[BEANFS_NDADDR];     // direct addr; offset in data blocks
-        unsigned long i_id_offset[BEANFS_NINDIR];     // first is 1st direct and second is 2th indirect
+        uint32_t i_db_offset[BEANFS_NDADDR];     // direct addr; offset in data blocks
+        uint32_t i_id_offset[BEANFS_NINDIR];     // first is 1st direct and second is 2th indirect
     } i_addr;
     
-    unsigned short inode_lock;                  // 0 : unlocked and 1 : locked
+    uint16_t inode_lock;                  // 0 : unlocked and 1 : locked
 };
 
 #define MAX_ENTRY_NAME 20
 struct beanfs_dir_entry {
     
-    unsigned long d_ino;                        // 0 offset of the inode
-    const unsigned short d_enlen;               // 4 entry length
+    uint32_t d_ino;                        // 0 offset of the inode
+    const uint16_t d_enlen;               // 4 entry length
     char d_name[MAX_ENTRY_NAME];                // 5 name_length
     char d_file_type;                           // 25 file type 'd' or 's' or '-'
 };                                              // 26 bytes
-
-/**********************************************************************************
  
- 
-
 #endif
